move to server utilities
